requirements.txt
# Date utilities
python-dateutil==2.8.2
pytz>=2024.2

# FastAPI and server
fastapi==0.103.2
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic
pydantic==1.10.13
pydantic-settings==0.2.5

# Environment
python-dotenv==1.0.0

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# AI - Groq
groq==0.4.1

# JWT Authentication
PyJWT==2.8.0

# RAG / Vector DB
chromadb==0.3.29
sentence-transformers==2.2.2

# LangChain
langchain==0.1.0
langchain-community==0.0.10

# Email parsing
beautifulsoup4==4.12.2

# Async support
httpx==0.25.2
aiofiles==23.2.1

# Frontend
streamlit==1.29.0
requests==2.31.0
dateparser==1.3.0